requirements.txt
Python==3.10.12
PyTorch==2.0.1
OpenCV==4.8.0
Numpy==1.25.2
Matplotlib==3.7.2
PyQT==5.15.7
Qt==5.15.2